backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biomed-public-app-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "openai>=1.109.0",
  "langgraph>=1.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
